text-to-image/requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
opencv-python
safetensors
Pillow